RAGMaster/langchain_env_requirements.txt
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
asyncer==0.0.2
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.2
bidict==0.22.1
build==1.0.3
cachetools==5.3.2
certifi==2023.11.17
chainlit==1.0.101
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.22
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
dataclasses-json==0.5.14
Deprecated==1.2.14
et-xmlfile==1.1.0
fastapi==0.100.1
fastapi-socketio==0.0.10
filelock==3.13.1
filetype==1.2.0
flatbuffers==23.5.26
frozenlist==1.4.1
fsspec==2023.12.2
google-auth==2.26.2
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.60.0
h11==0.14.0
httpcore==0.17.3
httptools==0.6.1
httpx==0.24.1
huggingface-hub==0.20.2
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
jsonpatch==1.33
jsonpointer==2.4
kubernetes==29.0.0
langchain==0.1.1
langchain-community==0.0.13
langchain-core==0.1.12
langsmith==0.0.82
Lazify==0.4.0
literalai==0.0.102
marshmallow==3.20.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.9
networkx==3.1
nltk==3.8.1
numpy==1.26.3
oauthlib==3.2.2
onnxruntime==1.16.3
openai==0.28.0
openpyxl==3.1.2
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-exporter-otlp-proto-http==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
overrides==7.4.0
packaging==23.2
pandas==2.1.4
pip==23.3.1
posthog==3.3.1
protobuf==4.25.2
pulsar-client==3.4.0
pyaml==23.12.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
pypdf==3.17.4
PyPika==0.48.9
pyproject_hooks==1.0.0
pyprojroot==0.3.0
pyreadline3==3.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.8.2
python-graphql-client==0.4.3
python-multipart==0.0.6
python-socketio==5.11.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.3.3
sentence-transformers==2.2.2
sentencepiece==0.1.99
setuptools==68.2.2
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.27.0
sympy==1.12
syncer==2.0.3
tenacity==8.2.3
tiktoken==0.5.2
tokenizers==0.15.0
tomli==2.0.1
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.9.0
typing-inspect==0.9.0
tzdata==2023.4
uptrace==1.22.0
urllib3==2.1.0
uvicorn==0.23.2
watchfiles==0.20.0
websocket-client==1.7.0
websockets==12.0
wheel==0.41.2
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.17.0
